move space station to V2
